Renamed ListItem to Week because that made more sense to have a Week object with a bunch of different days in it. It will still handling The reading the file though.
